+++ packages/errors/src/index.ts
@@ -3,3 +3,15 @@ export class GitError extends Error {
     super(`Error: Git - ${message}, exit code: ${exitCode}`);
   }
 }
+
+export class PreExitButNotInPreModeError extends Error {
+  constructor() {
+    super("pre mode cannot be exited when not in pre mode");
+  }
+}
+
+export class PreEnterButInPreModeError extends Error {
+  constructor() {
+    super("pre mode cannot be entered when in pre mode");
+  }
+}
