+++ packages/get-release-plan/src/index.ts
@@ -4,6 +4,7 @@ import getWorkspaces from "get-workspaces";
 import getDependentsgraph from "get-dependents-graph";
 import { read } from "@changesets/config";
 import { Config, ReleasePlan } from "@changesets/types";
+import { readPreState } from "@changesets/pre";
 
 export default async function getReleasePlan(
   cwd: string,
@@ -17,13 +18,20 @@ export default async function getReleasePlan(
 
   if (!workspaces)
     throw new Error(
-      "Could not resolve workspaes for current working directory"
+      "Could not resolve workspaces for current working directory"
     );
 
+  const preState = await readPreState(cwd);
   const dependentsGraph = await getDependentsgraph({ cwd });
   const readConfig = await read(cwd, workspaces);
   const config = passedConfig ? { ...readConfig, ...passedConfig } : readConfig;
   const changesets = await readChangesets(cwd, sinceMaster);
 
-  return assembleReleasePlan(changesets, workspaces, dependentsGraph, config);
+  return assembleReleasePlan(
+    changesets,
+    workspaces,
+    dependentsGraph,
+    config,
+    preState
+  );
 }
