+++ packages/pre/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@changesets/pre",
+  "version": "0.2.2",
+  "description": "Utils to make a Changesets repo enter and exit pre mode",
+  "main": "dist/pre.cjs.js",
+  "module": "dist/pre.esm.js",
+  "license": "MIT",
+  "repository": "https://github.com/changesets/changesets/tree/master/packages/pre",
+  "dependencies": {
+    "@babel/runtime": "^7.4.4",
+    "@changesets/errors": "^0.1.0"
+  },
+  "devDependencies": {
+    "fixturez": "^1.1.0",
+    "outdent": "^0.5.0"
+  }
+}
